requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via redis
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
contextvars==2.4
    # via -r requirements.in
deprecated==1.2.13
    # via limits
dnspython==2.3.0
    # via email-validator
email-validator==1.3.1
    # via fastapi
fastapi[all]==0.92.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.23.3
    # via fastapi
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   rfc3986
immutables==0.19
    # via contextvars
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
limits==2.8.0
    # via slowapi
markupsafe==2.1.2
    # via jinja2
orjson==3.8.6
    # via fastapi
packaging==22.0
    # via limits
pydantic==1.10.5
    # via fastapi
pyjwt==2.6.0
    # via -r requirements.in
python-dotenv==0.21.1
    # via uvicorn
python-multipart==0.0.5
    # via fastapi
pyyaml==6.0
    # via
    #   fastapi
    #   uvicorn
redis==4.5.1
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-multipart
slowapi==0.1.7
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.25.0
    # via fastapi
typing-extensions==4.5.0
    # via
    #   limits
    #   pydantic
ujson==5.7.0
    # via fastapi
uvicorn[standard]==0.20.0
    # via fastapi
watchfiles==0.18.1
    # via uvicorn
websockets==10.4
    # via uvicorn
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
